+++ CHANGELOG.md
@@ -38,6 +38,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Replaced Sphinx with Mkdocs for documentation generation.
 - *BREAKING CHANGE:* Renamed `KokoroDeviceNames` to `KokoroDeviceTypes`.
+- *BREAKING CHANGE:* `TTSAudioSpec.format` attribute changed to `mime_type`.
+- *BREAKING CHANGE:* `KokoroBackend` now audio output now conforms to the `audio/L16`
+  MIME type, which means it's output is now in big endian byte order instead of the
+  previous little endian.
 
 ### Deprecated